# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "TG Summarizer", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/network/test-proxy": {
      "post": {
        "tags": ["network"],
        "summary": "Api Test Proxy",
        "operationId": "network-api_test_proxy",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TestProxyRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Network-Api Test Proxy"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/api/v1/network/tor-status": {
      "get": {
        "tags": ["network"],
        "summary": "Api Tor Status",
        "operationId": "network-api_tor_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Network-Api Tor Status"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/network/tor-ip": {
      "get": {
        "tags": ["network"],
        "summary": "Api Tor Ip",
        "operationId": "network-api_tor_ip",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Network-Api Tor Ip"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/network/tor-restart": {
      "post": {
        "tags": ["network"],
        "summary": "Api Tor Restart",
        "operationId": "network-api_tor_restart",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Network-Api Tor Restart"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/network/tor-new-identity": {
      "post": {
        "tags": ["network"],
        "summary": "Api Tor New Identity",
        "operationId": "network-api_tor_new_identity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TorNewIdentityRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Network-Api Tor New Identity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
...
    "/api/v1/jobs/status": {
      "get": {
        "tags": ["jobs"],
        "summary": "Jobs Status",
        "operationId": "jobs-jobs_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Jobs-Jobs Status"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/jobs/runtime-config": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get Runtime Config",
        "operationId": "jobs-get_runtime_config",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeConfigResponse"
                }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/jobs/{job_id}/trigger": {
      "post": {
        "tags": ["jobs"],
        "summary": "Trigger Scheduler Job",
        "operationId": "jobs-trigger_scheduler_job",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Jobs-Trigger Scheduler Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "put": {
        "tags": ["jobs"],
        "summary": "Update Scheduler Job",
        "operationId": "jobs-update_scheduler_job",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Job Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateJobRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Jobs-Update Scheduler Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/sync": {
      "post": {
        "tags": ["jobs"],
        "summary": "Start Sync Job",
        "operationId": "jobs-start_sync_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/StartSyncJobRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartSyncJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "OAuth2PasswordBearer": [] }]
      }
    },
    "/api/v1/jobs/sync/{job_id}": {
      "get": {
        "tags": ["jobs"],
        "summary": "Get Sync Job Status",
        "operationId": "jobs-get_sync_job_status",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncJobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/sync/{job_id}/events": {
      "get": {
        "tags": ["jobs"],
        "summary": "Sync Job Events",
        "operationId": "jobs-sync_job_events",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/sync/{job_id}/cancel": {
      "post": {
        "tags": ["jobs"],
        "summary": "Cancel Sync Job",
        "operationId": "jobs-cancel_sync_job",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Job Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelSyncJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActiveSyncJobSummary": {
        "properties": {
          "jobId": { "type": "string", "title": "Jobid" },
          "status": { "type": "string", "title": "Status" },
          "source": { "type": "string", "title": "Source" },
          "channelCount": { "type": "integer", "title": "Channelcount" },
          "runningChannels": { "type": "integer", "title": "Runningchannels" },
          "pendingChannels": { "type": "integer", "title": "Pendingchannels" },
          "allowedConcurrency": {
            "type": "integer",
            "title": "Allowedconcurrency"
          },
          "concurrencyInUse": {
            "type": "integer",
            "title": "Concurrencyinuse"
          },
          "effectiveProxyCapacity": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Effectiveproxycapacity"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status",
          "source",
          "channelCount",
          "runningChannels",
          "pendingChannels",
          "allowedConcurrency",
          "concurrencyInUse"
        ],
        "title": "ActiveSyncJobSummary"
      },
...
      "CancelSyncJobResponse": {
        "properties": {
          "jobId": { "type": "string", "title": "Jobid" },
          "status": { "type": "string", "title": "Status" }
        },
        "type": "object",
        "required": ["jobId", "status"],
        "title": "CancelSyncJobResponse"
      },
...
      "GlobalStartTimeSnapshot": {
        "properties": {
          "mode": { "type": "string", "title": "Mode" },
          "value": {
            "anyOf": [
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" },
              { "type": "null" }
            ],
            "title": "Value"
          },
          "effectiveMs": { "type": "integer", "title": "Effectivems" }
        },
        "type": "object",
        "required": ["mode", "effectiveMs"],
        "title": "GlobalStartTimeSnapshot"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobsRuntimeSettings": {
        "properties": {
          "enabled": {
            "additionalProperties": { "type": "boolean" },
            "type": "object",
            "title": "Enabled"
          },
          "intervals": { "$ref": "#/components/schemas/JobIntervals" }
        },
        "type": "object",
        "required": ["enabled", "intervals"],
        "title": "JobsRuntimeSettings"
      },
...
      "NetworkRuntimeSettings": {
        "properties": {
          "proxyEnabled": {
            "type": "boolean",
            "title": "Proxyenabled",
            "default": false
          },
          "proxyUrls": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Proxyurls"
          },
          "proxyDefaultConcurrency": {
            "type": "integer",
            "title": "Proxydefaultconcurrency",
            "default": 1
          },
          "proxyConcurrencyOverrides": {
            "additionalProperties": { "type": "integer" },
            "type": "object",
            "title": "Proxyconcurrencyoverrides"
          },
          "effectiveProxyCapacity": {
            "type": "integer",
            "title": "Effectiveproxycapacity",
            "default": 0
          },
          "proxyLanes": {
            "items": { "$ref": "#/components/schemas/ProxyLaneSnapshot" },
            "type": "array",
            "title": "Proxylanes"
          },
          "torEnabled": {
            "type": "boolean",
            "title": "Torenabled",
            "default": false
          },
          "torMode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Tormode"
          },
          "torProxyUrls": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Torproxyurls"
          },
          "torRotationStrategy": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Torrotationstrategy"
          },
          "torControlEnabled": {
            "type": "boolean",
            "title": "Torcontrolenabled",
            "default": false
          },
          "torControlPort": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Torcontrolport"
          },
          "torAutoRotate": {
            "type": "boolean",
            "title": "Torautorotate",
            "default": false
          },
          "torRotationThreshold": {
            "type": "integer",
            "title": "Torrotationthreshold"
          },
          "envFallbackConfigured": {
            "type": "boolean",
            "title": "Envfallbackconfigured"
          },
          "usingEnvFallback": {
            "type": "boolean",
            "title": "Usingenvfallback"
          },
          "torAvailable": { "type": "boolean", "title": "Toravailable" },
          "torSocksProxy": { "type": "string", "title": "Torsocksproxy" },
          "effectiveTorEnabled": {
            "type": "boolean",
            "title": "Effectivetorenabled"
          },
          "resolvedProxyCount": {
            "type": "integer",
            "title": "Resolvedproxycount"
          },
          "resolvedProxies": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Resolvedproxies"
          },
          "fetchTimeoutSeconds": {
            "type": "number",
            "title": "Fetchtimeoutseconds"
          },
          "fetchRetries": { "type": "integer", "title": "Fetchretries" },
          "fetchInitialDelayMs": {
            "type": "integer",
            "title": "Fetchinitialdelayms"
          },
          "proxyCooldownMs": { "type": "integer", "title": "Proxycooldownms" },
          "telegramApiRetries": {
            "type": "integer",
            "title": "Telegramapiretries"
          },
          "telegramApiInitialDelayMs": {
            "type": "integer",
            "title": "Telegramapiinitialdelayms"
          }
        },
        "type": "object",
        "required": [
          "torRotationThreshold",
          "envFallbackConfigured",
          "usingEnvFallback",
          "torAvailable",
          "torSocksProxy",
          "effectiveTorEnabled",
          "resolvedProxyCount",
          "fetchTimeoutSeconds",
          "fetchRetries",
          "fetchInitialDelayMs",
          "proxyCooldownMs",
          "telegramApiRetries",
          "telegramApiInitialDelayMs"
        ],
        "title": "NetworkRuntimeSettings"
      },
...
      "RetentionRuntimeSettings": {
        "properties": {
          "postRetentionDays": {
            "type": "integer",
            "title": "Postretentiondays"
          },
          "logRetentionDays": { "type": "integer", "title": "Logretentiondays" }
        },
        "type": "object",
        "required": ["postRetentionDays", "logRetentionDays"],
        "title": "RetentionRuntimeSettings"
      },
      "RuntimeConfigResponse": {
        "properties": {
          "nowMs": { "type": "integer", "title": "Nowms" },
          "scrapeCutoffMs": { "type": "integer", "title": "Scrapecutoffms" },
          "effectiveGlobalStartTimeMs": {
            "type": "integer",
            "title": "Effectiveglobalstarttimems"
          },
          "globalStartTime": {
            "$ref": "#/components/schemas/GlobalStartTimeSnapshot"
          },
          "sync": { "$ref": "#/components/schemas/SyncRuntimeSettings" },
          "scraper": { "$ref": "#/components/schemas/ScraperRuntimeSettings" },
          "network": { "$ref": "#/components/schemas/NetworkRuntimeSettings" },
          "jobs": { "$ref": "#/components/schemas/JobsRuntimeSettings" },
          "retention": {
            "$ref": "#/components/schemas/RetentionRuntimeSettings"
          },
          "constants": { "$ref": "#/components/schemas/RuntimeConstants" },
          "activeSyncJob": {
            "anyOf": [
              { "$ref": "#/components/schemas/ActiveSyncJobSummary" },
              { "type": "null" }
            ]
          }
        },
        "type": "object",
        "required": [
          "nowMs",
          "scrapeCutoffMs",
          "effectiveGlobalStartTimeMs",
          "globalStartTime",
          "sync",
          "scraper",
          "network",
          "jobs",
          "retention",
          "constants"
        ],
        "title": "RuntimeConfigResponse"
      },
      "RuntimeConstants": {
        "properties": {
          "syncMaxRetries": { "type": "integer", "title": "Syncmaxretries" },
          "syncRetryBackoffBaseMs": {
            "type": "integer",
            "title": "Syncretrybackoffbasems"
          },
          "syncJobSseThrottleMs": {
            "type": "integer",
            "title": "Syncjobssethrottlems"
          },
          "syncJobPersistIntervalMs": {
            "type": "integer",
            "title": "Syncjobpersistintervalms"
          },
          "autoSyncPauseDurationMs": {
            "type": "integer",
            "title": "Autosyncpausedurationms"
          },
          "autoSyncFailureThresholdMin": {
            "type": "integer",
            "title": "Autosyncfailurethresholdmin"
          },
          "embeddingsChunkSize": {
            "type": "integer",
            "title": "Embeddingschunksize"
          },
          "embeddingsBackfillLimitDefault": {
            "type": "integer",
            "title": "Embeddingsbackfilllimitdefault"
          },
          "translationBatchLimit": {
            "type": "integer",
            "title": "Translationbatchlimit"
          },
          "translationBatchMaxChars": {
            "type": "integer",
            "title": "Translationbatchmaxchars"
          },
          "environment": { "type": "string", "title": "Environment" }
        },
        "type": "object",
        "required": [
          "syncMaxRetries",
          "syncRetryBackoffBaseMs",
          "syncJobSseThrottleMs",
          "syncJobPersistIntervalMs",
          "autoSyncPauseDurationMs",
          "autoSyncFailureThresholdMin",
          "embeddingsChunkSize",
          "embeddingsBackfillLimitDefault",
          "translationBatchLimit",
          "translationBatchMaxChars",
          "environment"
        ],
        "title": "RuntimeConstants"
      },
...
      "ScraperRuntimeSettings": {
        "properties": {
          "maxPostsPerChannel": {
            "type": "integer",
            "title": "Maxpostsperchannel"
          },
          "iterationLimit": { "type": "integer", "title": "Iterationlimit" }
        },
        "type": "object",
        "required": ["maxPostsPerChannel", "iterationLimit"],
        "title": "ScraperRuntimeSettings"
      },
      "StartSyncJobRequest": {
        "properties": {
          "channelIds": {
            "anyOf": [
              { "items": { "type": "string" }, "type": "array" },
              { "type": "null" }
            ],
            "title": "Channelids"
          },
          "source": { "type": "string", "title": "Source", "default": "Manual" }
        },
        "type": "object",
        "title": "StartSyncJobRequest"
      },
      "StartSyncJobResponse": {
        "properties": { "jobId": { "type": "string", "title": "Jobid" } },
        "type": "object",
        "required": ["jobId"],
        "title": "StartSyncJobResponse"
      },
...
      "SyncJobStatusResponse": {
        "properties": {
          "jobId": { "type": "string", "title": "Jobid" },
          "status": { "type": "string", "title": "Status" },
          "source": { "type": "string", "title": "Source" },
          "channels": {
            "items": { "$ref": "#/components/schemas/ChannelSyncProgress" },
            "type": "array",
            "title": "Channels"
          },
          "createdAt": { "type": "integer", "title": "Createdat" },
          "finishedAt": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Finishedat"
          }
        },
        "type": "object",
        "required": ["jobId", "status", "source", "channels", "createdAt"],
        "title": "SyncJobStatusResponse"
      },
      "SyncRuntimeSettings": {
        "properties": {
          "autoSyncEnabled": { "type": "boolean", "title": "Autosyncenabled" },
          "autoSyncInterval": {
            "type": "integer",
            "title": "Autosyncinterval"
          },
          "syncConcurrency": { "type": "integer", "title": "Syncconcurrency" },
          "consecutiveFailures": {
            "type": "integer",
            "title": "Consecutivefailures"
          },
          "autoSyncPauseUntil": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Autosyncpauseuntil"
          },
          "globalStartTimeMode": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Globalstarttimemode"
          },
          "globalStartTimeValue": {
            "anyOf": [
              { "type": "string" },
              { "type": "integer" },
              { "type": "number" },
              { "type": "null" }
            ],
            "title": "Globalstarttimevalue"
          }
        },
        "type": "object",
        "required": [
          "autoSyncEnabled",
          "autoSyncInterval",
          "syncConcurrency",
          "consecutiveFailures"
        ],
        "title": "SyncRuntimeSettings"
      },
      "TestProxyRequest": {
        "properties": { "proxyUrl": { "type": "string", "title": "Proxyurl" } },
        "type": "object",
        "required": ["proxyUrl"],
        "title": "TestProxyRequest"
      },
...
      "TorNewIdentityRequest": {
        "properties": {
          "port": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Port"
          }
        },
        "type": "object",
        "title": "TorNewIdentityRequest"
      },
...
      "UpdateJobRequest": {
        "properties": { "enabled": { "type": "boolean", "title": "Enabled" } },
        "type": "object",
        "required": ["enabled"],
        "title": "UpdateJobRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" },
          "input": { "title": "Input" },
          "ctx": { "type": "object", "title": "Context" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": { "scopes": {}, "tokenUrl": "/api/v1/login/access-token" }
        }
      }
    }
  }
}